Hotfix for logdet issue, and glm back to maxiter
There was a sever issue with the logdet calculation in solve_posdef, where the
sign was flipped. This explains most of the bad performance with revrand SLM.
Also it is easier to specify the number of iterations for SGD than number of
passes. So all algorihtms using sgd now refer to maxiter rather than passes.